Announce allowed purposes in gateway
This change signals to relays (see payjoin/ohttp-relay#58) that even if
the gateway is not known to the relay, requests are still allowed for
the purposes of BIP 77.

Because receivers choose the directory (which is its own gateway), but
senders choose the relay, before introducing this functionality it this
would have been the responsibility of the sender (or more realistically
their wallet vendor) to choose an appropriate relay, one that accepts
requests on behalf of the receiver's chosen directory.

Rationale for adding a UUID: although arguably 64 bits of randomness is
more than enough, and even just the string "BIP 77" is likely to be
unique and unambiguous, adding a UUID practically ensures an opt-in is
really an opt-in as 128 random bits would not be repeated except
intentionally.

Rationale for format: we just need a set of opaque identifiers, encoding
the same way as TLS ALPN does makes sense since that's well specified,
binary safe, easy to parse and places readonable limits on string
lengths.